+++ src/Console/ChainOutput.php
@@ -9,19 +9,29 @@ use Symfony\Component\Console\Output\OutputInterface;
 
 final readonly class ChainOutput implements OutputInterface
 {
-	private ConsoleOutput $consoleOutput;
+	private OutputInterface $mainOutput;
 	private BufferedOutput $bufferedOutput;
 
 	/**
+	 * @param 8|16|32|64|128|256 $verbosity
 	 */
 	public function __construct (
 		int $verbosity = self::VERBOSITY_NORMAL,
 		bool $decorated = true,
 		?OutputFormatterInterface $formatter = null,
+		?OutputInterface $mainOutput = null,
 	)
 	{
 		$this->bufferedOutput = new BufferedOutput($verbosity, $decorated, $formatter);
-		$this->consoleOutput = new ConsoleOutput($verbosity, $decorated, $formatter);
+
+		$this->mainOutput = $mainOutput ?? new ConsoleOutput();
+		$this->mainOutput->setDecorated($decorated);
+		$this->mainOutput->setVerbosity($verbosity);
+
+		if (null !== $formatter)
+		{
+			$this->mainOutput->setFormatter($formatter);
+		}
 	}
 
 	/**
@@ -31,7 +41,7 @@ final readonly class ChainOutput implements OutputInterface
 	public function write (iterable|string $messages, bool $newline = false, int $options = 0) : void
 	{
 		$this->bufferedOutput->write($messages, $newline, $options);
-		$this->consoleOutput->write($messages, $newline, $options);
+		$this->mainOutput->write($messages, $newline, $options);
 	}
 
 	/**
@@ -41,7 +51,7 @@ final readonly class ChainOutput implements OutputInterface
 	public function writeln (iterable|string $messages, int $options = 0) : void
 	{
 		$this->bufferedOutput->writeln($messages, $options);
-		$this->consoleOutput->writeln($messages, $options);
+		$this->mainOutput->writeln($messages, $options);
 	}
 
 	/**
@@ -51,7 +61,7 @@ final readonly class ChainOutput implements OutputInterface
 	public function setVerbosity (int $level) : void
 	{
 		$this->bufferedOutput->setVerbosity($level);
-		$this->consoleOutput->setVerbosity($level);
+		$this->mainOutput->setVerbosity($level);
 	}
 
 	/**
@@ -113,8 +123,8 @@ final readonly class ChainOutput implements OutputInterface
 	#[\Override]
 	public function setDecorated (bool $decorated) : void
 	{
-		$this->bufferedOutput->setDecorated(true);
-		$this->consoleOutput->setDecorated(true);
+		$this->bufferedOutput->setDecorated($decorated);
+		$this->mainOutput->setDecorated($decorated);
 	}
 
 	/**
@@ -132,7 +142,7 @@ final readonly class ChainOutput implements OutputInterface
 	public function setFormatter (OutputFormatterInterface $formatter) : void
 	{
 		$this->bufferedOutput->setFormatter($formatter);
-		$this->consoleOutput->setFormatter($formatter);
+		$this->mainOutput->setFormatter($formatter);
 	}
 
 	/**