+++ src/Entity/TaskLog.php
@@ -24,6 +24,8 @@ use function Symfony\Component\Clock\now;
  */
 #[ORM\Entity]
 #[ORM\Table(name: "task_manager_tasks")]
+#[ORM\Index(columns: ["time_queued"], name: "idx_task_manager_tasks_time_queued")]
+#[ORM\Index(columns: ["taskId"], name: "idx_task_manager_task_id")]
 class TaskLog
 {
 	/**
@@ -194,18 +196,16 @@ class TaskLog
 	}
 
 	/**
-	 * Returns the unserialized cached task object
+	 * @deprecated Use TaskDetailsNormalizer::deserializeTask() instead. Will be removed in 4.0.
+	 *
+	 * @todo Remove in 4.0.
 	 */
-	public function getTaskObject () : ?object
+	public function getTaskObject () : null
 	{
-		$task = $this->getTaskDetails()["task"] ?? null;
-		$unserialized = \is_string($task)
-			? unserialize($task)
-			: null;
+		// @phpstan-ignore-next-line todoBy.sfDeprecation
+		trigger_deprecation("21torr/task-manager", "3.2.5", "TaskLog::getTaskObject() is deprecated, use TaskDetailsNormalizer::deserializeTask() instead.");
 
-		return \is_object($unserialized)
-			? $unserialized
-			: null;
+		return null;
 	}
 
 	/**