+++ src/Normalizer/TaskDetailsNormalizer.php
@@ -2,9 +2,13 @@
 
 namespace Torr\TaskManager\Normalizer;
 
+use Psr\Log\LoggerInterface;
 use Symfony\Component\Messenger\Envelope;
 use Symfony\Component\Messenger\Stamp\HandledStamp;
 use Symfony\Component\Messenger\Stamp\ReceivedStamp;
+use Symfony\Component\Serializer\Encoder\JsonEncoder;
+use Symfony\Component\Serializer\Exception\ExceptionInterface as SerializerException;
+use Symfony\Component\Serializer\SerializerInterface;
 use Torr\TaskManager\Entity\TaskLog;
 use Torr\TaskManager\Task\Task;
 
@@ -13,6 +17,11 @@ use Torr\TaskManager\Task\Task;
  */
 final readonly class TaskDetailsNormalizer
 {
+	public function __construct (
+		private SerializerInterface $serializer,
+		private LoggerInterface $logger,
+	) {}
+
 	/**
 	 * @return TaskDetails
 	 */
@@ -29,9 +38,40 @@ final readonly class TaskDetailsNormalizer
 		if ($task instanceof Task)
 		{
 			$details["label"] = $task->getMetaData()->label;
-			$details["task"] = serialize($task);
+			$details["task"] = $this->serializer->serialize($task, JsonEncoder::FORMAT);
 		}
 
 		return $details;
+	}
+
+	/**
+	 * Deserializes the task object stored in the given log entry.
+	 */
+	public function deserializeTask (TaskLog $log) : ?Task
+	{
+		$serialized = $log->getTaskDetails()["task"] ?? null;
+
+		if (!\is_string($serialized))
+		{
+			return null;
+		}
+
+		try
+		{
+			$task = $this->serializer->deserialize($serialized, $log->taskClass, JsonEncoder::FORMAT);
+
+			return $task instanceof Task ? $task : null;
+		}
+		catch (SerializerException $exception)
+		{
+			$this->logger->error("Failed to deserialize task of class '{taskClass}' from log entry #{logId}: {message}", [
+				"taskClass" => $log->taskClass,
+				"logId" => $log->id,
+				"message" => $exception->getMessage(),
+				"exception" => $exception,
+			]);
+
+			return null;
+		}
 	}
 }